+++ Editor/Drawing/Drawing.cs
@@ -1,15 +1,39 @@
-﻿using System;
+using System;
 using System.Drawing;
 using System.Drawing.Imaging;
 using System.Runtime.InteropServices;
 using System.Threading.Tasks;
 
-namespace WalkerSim
+namespace WalkerSim.Editor.Drawing
 {
-    internal static class ImageUtils
+    internal class SystemDrawingImpl : WalkerSim.Drawing.IBitmap
     {
-        public static void RemoveTransparency(Bitmap img)
+        public int Width => Inner?.Width ?? 0;
+
+        public int Height => Inner?.Height ?? 0;
+
+        public readonly System.Drawing.Bitmap Inner;
+
+        public SystemDrawingImpl(Bitmap bitmap)
         {
+            Inner = bitmap;
+        }
+
+        public void Dispose()
+        {
+            Inner?.Dispose();
+        }
+
+        public WalkerSim.Drawing.Color GetPixel(int x, int y)
+        {
+            var src = Inner.GetPixel(x, y);
+            return new WalkerSim.Drawing.Color(
+                src.R, src.G, src.B, src.A);
+        }
+
+        public void RemoveTransparency()
+        {
+            var img = Inner;
             var rect = new Rectangle(0, 0, img.Width, img.Height);
             var bmpData = img.LockBits(rect, ImageLockMode.ReadWrite, img.PixelFormat);
 
@@ -53,6 +77,22 @@ namespace WalkerSim
 
             Marshal.Copy(pixels, 0, ptrFirstPixel, pixels.Length);
             img.UnlockBits(bmpData);
+        }
+    }
+
+    internal class ImageLoader : WalkerSim.Drawing.IImageLoader
+    {
+        public WalkerSim.Drawing.IBitmap CreateBitmap(WalkerSim.Drawing.IBitmap src, int width, int height)
+        {
+            var bitmapSrc = src as SystemDrawingImpl;
+            return new SystemDrawingImpl(new System.Drawing.Bitmap(bitmapSrc.Inner, width, height));
+        }
+
+        public WalkerSim.Drawing.IBitmap LoadFromFile(string filePath)
+        {
+            var bitmapSrc = System.Drawing.Image.FromFile(filePath);
+
+            return new SystemDrawingImpl((Bitmap)bitmapSrc);
         }
     }
 }